major refactor of organization of ogv dashboard code
