+++ package.json
@@ -51,6 +51,7 @@
     "sanitize-html": "^2.7.3",
     "sentry": "^0.1.2",
     "sharp": "^0.31.2",
+    "tailwind-merge": "^1.9.0",
     "tailwindcss": "^3.2.4",
     "use-analytics": "^0.0.5"
   },
