+++ pages/ogv-dashboard/chart-configs/index.ts
@@ -1,0 +1,2 @@
+export { default as lineOptions } from "./lineOptions";
+export { default as doughnutOptions } from "./doughnutOptions";
