+++ pages/ogv-dashboard/components/index.ts
@@ -1,0 +1,2 @@
+export { default as Section } from "./Section";
+export { default as ChartButton } from "./ChartButton";
