+++ pages/ogv-dashboard/constants.ts
@@ -1,0 +1,5 @@
+export const smSize = 640;
+
+export const gradientStart = "#8C66FC";
+
+export const gradientEnd = "#0274F1";
