+++ pages/ogv-dashboard/data/index.ts
@@ -1,0 +1,2 @@
+export { default as nonCirculatingSupply } from "./nonCirculatingSupply";
+export { default as doughnutData } from "./doughnutData";
