+++ pages/ogv-dashboard/data/nonCirculatingSupply.ts
@@ -1,0 +1,59 @@
+export const nonCirculatingSupply = [
+  {
+    address: "0xbe2AB3d3d8F6a32b96414ebbd865dBD276d3d899",
+    internalLabel: "5 of 8",
+    publicLabel: "Foundation Reserves",
+  },
+  {
+    address: "0xcaa5eF7AbC36D5E5a3E4d7930DCfF3226617A167",
+    internalLabel: "Old T3 Team Hot Wallet",
+    publicLabel: "Team Distribution",
+  },
+  {
+    address: "0x2EaE0CaE2323167ABF78462e0C0686865c67a655",
+    internalLabel: "New T3 Team Hot Wallet",
+    publicLabel: "Team Distribution",
+  },
+  {
+    address: "0x3Da5045699802Ea1fCc60130dEDEa67139C5b8C0",
+    internalLabel: "Old T3 Investor Hot Wallet",
+    publicLabel: "Investor Distribution",
+  },
+  {
+    address: "0xFE730B3cf80cA7B31905f70241F7C786BAF443E3",
+    internalLabel: "New T3 Investor Hot Wallet",
+    publicLabel: "Investor Distribution",
+  },
+  {
+    address: "0x12D7EF3C933D091210cD931224Ead45D9cFdDdE0",
+    internalLabel: "Lukewarm",
+    publicLabel: "Distribution Staging",
+  },
+  {
+    address: "0x0C4576Ca1c365868E162554AF8e385dc3e7C66D9",
+    internalLabel: "Staked OGV",
+    publicLabel: "Staked OGV",
+  },
+  {
+    address: "0x7aE2334f12a449895AD21d4c255D9DE194fe986f",
+    internalLabel: "OGV Claims",
+    publicLabel: "OGV Claims",
+  },
+  {
+    address: "0xD667091c2d1DCc8620f4eaEA254CdFB0a176718D",
+    internalLabel: "veOGV Claims",
+    publicLabel: "veOGV Claims",
+  },
+  {
+    address: "0x8ac3b96d118288427055ae7f62e407fC7c482F57",
+    internalLabel: "Custodian for Brave Endeavors",
+    publicLabel: "Custodian",
+  },
+  {
+    address: "0xA2Cc2eAE69cBf04a3D5660bc3E689B035324Fc3F",
+    internalLabel: "Custodian for Limitless Alpha",
+    publicLabel: "Custodian",
+  },
+];
+
+export default nonCirculatingSupply;
