+++ pages/ogv-dashboard/sections/index.ts
@@ -1,0 +1,7 @@
+export { default as Heading } from "./Heading";
+export { default as OgvPriceStats } from "./OgvPriceStats";
+export { default as OgvPriceChart } from "./OgvPriceChart";
+export { default as StakingBanner } from "./StakingBanner";
+export { default as OgvStakingStats } from "./OgvStakingStats";
+export { default as AllocationDistribution } from "./AllocationDistribution";
+export { default as TopExchanges } from "./TopExchanges";
