+++ pages/ogv-dashboard/types/ChartTime.ts
@@ -1,0 +1,8 @@
+export enum ChartTime {
+  ONE_DAY = 1,
+  SEVEN_DAY = 7,
+  THIRTY_DAY = 30,
+  ONE_YEAR = 365,
+}
+
+export default ChartTime;
