+++ pages/ogv-dashboard/types/ChartType.ts
@@ -1,0 +1,6 @@
+export enum ChartType {
+  Price,
+  Market_Cap,
+}
+
+export default ChartType;
