+++ pages/ogv-dashboard/types/GetStorageAtResponse.ts
@@ -1,0 +1,7 @@
+export interface GetStorageAtResponse {
+  jsonrpc: string;
+  id: number;
+  result: string;
+}
+
+export default GetStorageAtResponse;
