+++ pages/ogv-dashboard/types/index.ts
@@ -1,0 +1,6 @@
+export type { default as Link } from "./Link";
+export type { default as DashProps } from "./DashProps";
+export type { default as GetStorageAtResponse } from "./GetStorageAtResponse";
+export type { default as GetBlockByNumberResponse } from "./GetBlockByNumberResponse";
+export { default as ChartTime } from "./ChartTime";
+export { default as ChartType } from "./ChartType";
